Se termina de migrar la views de la app loginAuth0 a la app accounts
