+++ ProjectManager/urls.py
@@ -6,6 +6,6 @@ from account import urls as account_urls
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include(loginauth_urls)),
     path('', include(account_urls)),
+    path('', include(loginauth_urls)),
 ]
